requirements.txt
absl-py==0.11.0
aioredis==1.3.1
altgraph==0.17
argon2-cffi==20.1.0
asgi-redis==1.4.3
asgiref==3.2.10
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
constantly==15.1.0
convertdate==2.3.2
cryptography==3.4.6
cycler==0.10.0
daphne==3.0.1
decorator==4.4.2
defusedxml==0.7.0
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-eventstream==4.0.0
django-grip==3.0.0
django-jsonfield==1.4.1
django-multiforloop==0.2.1
django-rename-app==0.1.3
django-workers==0.2.0
entrypoints==0.3
Faker==8.0.0
folium==0.12.1
future==0.18.2
geojson==2.5.0
gripcontrol==4.0.0
hijri-converter==2.1.1
hiredis==1.1.0
holidays==0.11.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
ipykernel==5.5.0
ipyleaflet==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
Js2Py==0.71
jsons==1.4.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
msgpack==1.0.2
msgpack-python==0.5.6
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
openrouteservice==2.3.3
ortools==8.1.8487
OSMPythonTools==0.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.1
pubcontrol==3.0.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyjsparser==2.7.1
PyJWT==1.7.1
PyMeeus==0.5.11
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.17.3
pytest==6.2.3
pytest-sugar==0.9.4
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
redis==2.10.6
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
simpy==4.0.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
traittypes==0.2.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typish==1.9.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xarray==0.17.0
zope.interface==5.2.0
